Move TurtleDiagnostic interface to tree-sitter module
